requirements.txt
attrs==22.2.0
build==0.10.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
coverage==7.2.1
crashtest==0.4.1
distlib==0.3.6
dulwich==0.21.3
filelock==3.10.0
funkify==0.4.4
google-api-core==2.11.0
google-api-python-client==2.81.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
html5lib==1.1
httplib2==0.21.0
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
installer==0.6.0
isodate==0.6.1
jaraco.classes==3.2.3
jsonschema==4.17.3
keyring==23.13.1
linecache2==1.0.0
lockfile==0.12.2
more-itertools==9.1.0
msgpack==1.0.5
package-name==0.1
packaging==23.0
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==2.6.2
pluggy==1.0.0
poetry==1.4.0
poetry-core==1.5.1
poetry-plugin-export==1.3.0
protobuf==4.22.1
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==7.2.2
pytest-cov==4.0.0
rapidfuzz==2.13.7
requests==2.28.2
requests-toolbelt==0.10.1
requires==0.10.3
rsa==4.9
shellingham==1.5.0.post1
six==1.16.0
tomlkit==0.11.6
traceback2==1.4.0
trove-classifiers==2023.3.9
typing_extensions==4.5.0
unittest2==1.1.0
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.21.0
webencodings==0.5.1
xattr==0.10.1
xtyping==0.6.2
zipp==3.15.0
